moved PathMetadata to sciond
